New pathing system
Whole new AI pathfinding system using a bunch of custom code in AI manager, Enemy spawner, and the wall spawner.  Enemies current location and chosen path constantly get applied extra cost, diverting enemies away from overlapping. Disabled enemy v enemy collisions to prevent any rare chance to get stuck

+++ Assets/ScriptableObjects/MeleeSettings/MeleeSettings.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public class MeleeSettings : ScriptableObject
 {
     public bool destroyAfterAttack = false;
-    public int damage = 2;
-    public int fireRate = 1;
+    public float damage = 2;
+    public float fireRate = 1;
 }
